128-bit integers

+++ fixedpt.h
@@ -10,17 +10,17 @@
 
 namespace fixed {
 
-enum Num : long long { f0 = 0 };
+enum Num : __int128 { f0 = 0 };
 
 namespace impl {
 extern unsigned negExponent;
-extern unsigned long long scale;
+extern unsigned __int128 scale;
 }  // namespace impl
 extern void init(unsigned negExponent);
 extern Num lowest;
 extern Num min;
 extern Num max;
-inline Num toNum(long long value) {
+inline Num toNum(__int128 value) {
   return static_cast<Num>(value * impl::scale);
 }
 inline Num toNum(int value) { return static_cast<Num>(value * impl::scale); }
@@ -35,48 +35,49 @@ extern std::string toString(Num a);
 
 template <typename T>
 inline Num operator+(Num a, T b) {
-  return static_cast<Num>((long long)a + b * impl::scale);
+  return static_cast<Num>(static_cast<__int128>(a) + b * impl::scale);
 }
 template <>
 inline Num operator+(Num a, Num b) {
-  return static_cast<Num>((long long)a + (long long)b);
+  return static_cast<Num>(static_cast<__int128>(a) + static_cast<__int128>(b));
 }
 
 template <typename T>
 inline Num operator-(Num a, T b) {
-  return static_cast<Num>((long long)a - b * impl::scale);
+  return static_cast<Num>(static_cast<__int128>(a) - b * impl::scale);
 }
 template <>
 inline Num operator-(Num a, Num b) {
-  return static_cast<Num>((long long)a - (long long)b);
+  return static_cast<Num>(static_cast<__int128>(a) - static_cast<__int128>(b));
 }
 
 template <typename T>
 inline Num operator*(Num a, T b) {
-  return static_cast<Num>((long long)a * b);
+  return static_cast<Num>(static_cast<__int128>(a) * b);
 }
 template <>
 inline Num operator*(Num a, Num b) {
-  return static_cast<Num>((long double)(long long)a * (long long)b /
-                          impl::scale);
+  return static_cast<Num>((long double)static_cast<__int128>(a) *
+                          static_cast<__int128>(b) / impl::scale);
 }
 
 template <typename T>
 inline Num operator/(Num a, T b) {
-  return static_cast<Num>((long long)a / b);
+  return static_cast<Num>(static_cast<__int128>(a) / b);
 }
 template <>
 inline Num operator/(Num a, Num b) {
-  return static_cast<Num>(impl::scale * (long long)a / (long long)b);
+  return static_cast<Num>(impl::scale * static_cast<__int128>(a) /
+                          static_cast<__int128>(b));
 }
 
 template <typename T>
 inline bool operator>(Num a, T b) {
-  return (long long)a > b * impl::scale;
+  return static_cast<__int128>(a) > b * impl::scale;
 }
 template <>
 inline bool operator>(Num a, Num b) {
-  return (long long)a > (long long)b;
+  return static_cast<__int128>(a) > static_cast<__int128>(b);
 }
 extern std::ostream &operator<<(std::ostream &, Num);
